refactor: make store
